Update dunstrc to be compatible with a new version (+improvements)
Gets rid of `geometry` and uses the new `width height origin offset`
In changing this, I've made the notification offset be on a normal monitor the same from both the right side and the top side + topbar
I've also changed the default action on mouse_left_click to both do the embedded action and close the clicked notification.
For the time being, a new Volume section has been added to give the volume notification just a 1 second timeout. It's already set to accept a pull request from instantassist to also show the volume percentage. Maybe later the `summary` section could be replaced with tags or something of that nature.
Actually changed the progress bar to be the system color.